feat: expose `reduceConfigs` helper
