+++ packages/core/src/provider/rspackConfig.ts
@@ -6,12 +6,12 @@ import {
   type RspackConfig,
   castArray,
   getNodeEnv,
-  reduceConfigsAsyncWithContext,
 } from '@rsbuild/shared';
 import { rspack } from '@rspack/core';
 import { chainToConfig, modifyBundlerChain } from '../configChain';
 import { logger } from '../logger';
 import { getHTMLPlugin } from '../pluginHelper';
+import { reduceConfigsAsyncWithContext } from '../reduceConfigs';
 import type { InternalContext } from '../types';
 
 async function modifyRspackConfig(
