+++ packages/shared/src/types/utils.ts
@@ -17,3 +17,17 @@ export type CompilerTapFn<
 > = {
   tap: (name: string, cb: CallBack) => void;
 };
+
+export type ConfigChain<T> = OneOrMany<T | ((config: T) => T | void)>;
+
+export type ConfigChainWithContext<T, Ctx> = OneOrMany<
+  T | ((config: T, ctx: Ctx) => T | void)
+>;
+
+export type ConfigChainAsyncWithContext<T, Ctx> = OneOrMany<
+  T | ((config: T, ctx: Ctx) => MaybePromise<T | void>)
+>;
+
+export type ConfigChainMergeContext<T, Ctx> = OneOrMany<
+  T | ((merged: { value: T } & Ctx) => T | void)
+>;
